Add an AST pretty-printer
The pretty-printer formats code using a COBOL-like language that tries
to stay as compatible as possible with free-form COBOL. It is currently
not configurable (e.g. to output for different dialects).

Note that some issues are known regarding nested indexing (or indexing
within a function call), where either too much or not enough parentheses
may be added. This will be investigated and fixed in a separate PR.

The pretty-printer can be used with the `pp` command by passing the
`--parse` argument. To test the parser, the `--check` argument can be
passed to the `pp` command, which will check that re-parsing the
pretty-printed code gives back the same AST as the original.

The new `reparse` test checks the testsuite in the same way described
above.