Fix wrong position lookup upon rewind to the first line of input
This fixes a bug introduced in 86e6cca6 by an uncareful optimization
of the computation.  Essentially, the meaning of `Failure` and
`Invalid_argument` exception in `List.nth` needed to be reversed.  The
included changes fix and improve the affected portion of code, and add
some new (high-level) tests to better exercise the rewinding code.