Use minimal diffs for the indenter
Internally, the indenter computes a diff (in terms of offsets) for each
line that needs to change. Then, all the diffs are applied to the
document, and a single edit that rewrites the full document is sent to
the editor.

This PR changes the behavior to send small text edits that either add or
remove whitespace instead.