Use per-project file for indent configuration
The indent configuration is now exposed by the `Cobol_indent` module and
loaded from the `superbol.toml` file. For the sake of simplicity, this
does not change the internal representation of the indent configuration,
even though it requires us to expose the configuration as a hash table
in `Indent_type.ml`. Be responsible.

Fixes #46