Look for bundled superbol-free executable
This patch makes the VS Code extension look for a bundled executable in
the following order (and with a `.exe` suffix on Windows):

 - `_dist/superbol-free-${platform}-${arch}`
 - `_dist/superbol-free-${platform}`
 - `_dist/superbol-free`

So for instance on x64 Linux we will look for:

 - `_dist/superbol-free-linux-x64`
 - `_dist/superbol-free-linux`
 - `_dist/superbol-free`

On recent Macs we will look for:

 - `_dist/superbol-free-darwin-arm64`
 - `_dist/superbol-free-darwin`
 - `_dist/superbol-free`

And on Windows we will look for

 - `_dist/superbol-free-win32-x64.exe`
 - `_dist/superbol-free-win32.exe`
 - `_dist/superbol-free.exe`

This will also allow tackling #76 once we actually put the executables
in the extension. It also allows some flexibility between making "slim"
per-platform extensions or "fat" multiplatform extensions with many
executables.