Make the LSP server use user and/or workspace configurations
Note: the behavior with these changes will very likely be very messy
to maintain, as it essentially synchronizes `superbol.toml` with LSP
client settings.  In addition, it does not make much sense and makes
putting configurations under source control quite brittle.

To avoid this, the plan is to remove project-related configurations
from client settings, and better integrate editions of `superbol.toml`
with a JSON schema for this file and suggestions to use an extension
dedicated to TOML files.