New alternative, simpler indenter
This indenter is much simpler than the previous one. It uses a
per-division automaton working on a list of all the tokens in the
file. It covers most complex examples.

For now, it is configured using a .superbol-indent file that can be
contained in any parent directory of the file being indented.  The
--gen-config argument of "superbol-free indent file" can be used to
generate such a file in the current directory.

Also, the cobol-superbol-indent.el file is added for Emacs indentation
based on SuperBOL