Collect `EXEC(UTE)`/`END-EXEC` blocks in the pre-processor journal
Use entries from this journal to generate semantics tokens for
preamble and postamble locations only (instead of for the full text of
the blocks before this change).

Also includes a fix to properly recognize such blocks even when the
text that follows them is not (directly or indirectly) followed by a
period.